+++ doc/Changelog.md
@@ -2,6 +2,7 @@
 
 # git master
 
+* Added '--lb-logfile' command line option
 * Removed obsolete SKV backend for PersistentMap
 * [266](https://github.com/Eyescale/Lunchbox/pull/266):
   Add PersistentMap::getValues and PersistentMap::takeValues